requirements.txt
# UyKelishuv Bot - Minimal Requirements
python-telegram-bot==20.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Testing Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
httpx>=0.24.0